added a constant weight to the basic set and refactored constants into a single module
