+++ acceptance/test_mortality.py
@@ -8,7 +8,7 @@ from db_queries import get_envelope, get_outputs
 
 from cascade.executor.epiviz_runner import add_settings_to_execution_context
 from cascade.input_data.db.configuration import load_settings
-from cascade.dismod.db.metadata import IntegrandEnum
+from cascade.dismod.constants import IntegrandEnum
 from cascade.input_data.db.demographics import age_groups_to_ranges
 from cascade.dismod.db.wrapper import DismodFile, _get_engine
 from cascade.testing_utilities import make_execution_context
