+++ src/cascade/core/context.py
@@ -1,7 +1,7 @@
 from cascade.core.parameters import ParameterProperty
 from cascade.core.input_data import InputData
 from cascade.model.rates import Rate
-from cascade.dismod.db.metadata import IntegrandEnum
+from cascade.dismod.constants import IntegrandEnum
 
 from cascade.core.log import getLoggers
 CODELOG, MATHLOG = getLoggers(__name__)
