+++ src/cascade/executor/no_covariate_main.py
@@ -35,7 +35,7 @@ from cascade.executor.argument_parser import DMArgumentParser
 import cascade.input_data.db.bundle
 from cascade.input_data.configuration.builder import build_constraint
 from cascade.testing_utilities import make_execution_context
-from cascade.dismod.db.metadata import IntegrandEnum
+from cascade.dismod.constants import IntegrandEnum
 from cascade.dismod.db.wrapper import _get_engine
 from cascade.core.context import ModelContext
 from cascade.dismod.serialize import model_to_dismod_file
