+++ tests/input_data/test_emr.py
@@ -5,7 +5,7 @@ import numpy as np
 
 from pandas.testing import assert_frame_equal
 
-from cascade.dismod.db.metadata import DensityEnum
+from cascade.dismod.constants import DensityEnum
 from cascade.testing_utilities import make_execution_context
 from cascade.stats import meas_bounds_to_stdev
